Fix up tests a bit more
