+++ src/palace/manager/resources/opds2_schema/cached/drafts.opds.io/schema_acquisition-object.schema.json
@@ -1,0 +1,20 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$id": "https://drafts.opds.io/schema/acquisition-object.schema.json",
+  "title": "OPDS Acquisition Object",
+  "type": "object",
+  "properties": {
+    "type": {
+      "type": "string"
+    },
+    "child": {
+      "type": "array",
+      "items": {
+        "$ref": "acquisition-object.schema.json"
+      }
+    }
+  },
+  "required": [
+    "type"
+  ]
+}
