+++ src/palace/manager/resources/opds2_schema/cached/readium.org/webpub-manifest_schema_experimental_presentation_properties.schema.json
@@ -1,0 +1,31 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$id": "https://readium.org/webpub-manifest/schema/experimental/presentation/properties.schema.json",
+  "title": "Presentation Hints - Link Properties",
+  "type": "object",
+  "properties": {
+    "clipped": {
+      "description": "Specifies whether or not the parts of a linked resource that flow out of the viewport are clipped.",
+      "type": "boolean"
+    },
+    "fit": {
+      "description": "Specifies constraints for the presentation of a linked resource within the viewport.",
+      "type": "string",
+      "enum": [
+        "contain",
+        "cover",
+        "width",
+        "height"
+      ]
+    },
+    "orientation": {
+      "description": "Suggested orientation for the device when displaying the linked resource.",
+      "type": "string",
+      "enum": [
+        "auto",
+        "landscape",
+        "portrait"
+      ]
+    }
+  }
+}
