+++ src/palace/manager/resources/opds2_schema/cached/readium.org/webpub-manifest_schema_extensions_divina_properties.schema.json
@@ -1,0 +1,13 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$id": "https://readium.org/webpub-manifest/schema/extensions/divina/properties.schema.json",
+  "title": "Divina Profile - Link Properties",
+  "type": "object",
+  "properties": {
+    "break-scroll-before": {
+      "description": "Specifies that an item in the reading order should break the current continuous scroll and start a new one.",
+      "type": "boolean",
+      "default": false
+    }
+  }
+}
