+++ src/palace/manager/resources/opds2_schema/cached/readium.org/webpub-manifest_schema_extensions_epub_properties.schema.json
@@ -1,0 +1,32 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$id": "https://readium.org/webpub-manifest/schema/extensions/epub/properties.schema.json",
+  "title": "EPUB Profile - Link Properties",
+  "type": "object",
+  "properties": {
+    "contains": {
+      "description": "Identifies content contained in the linked resource, that cannot be strictly identified using a media type.",
+      "type": "array",
+      "items": {
+        "type": "string",
+        "enum": [
+          "mathml",
+          "onix",
+          "remote-resources",
+          "js",
+          "svg",
+          "xmp"
+        ]
+      },
+      "uniqueItems": true
+    },
+    "layout": {
+      "description": "Hints how the layout of a specific resource should be presented",
+      "type": "string",
+      "enum": [
+        "fixed",
+        "reflowable"
+      ]
+    }
+  }
+}
