+++ tests/manager/core/test_opds2_import.py
@@ -30,7 +30,7 @@ from palace.manager.sqlalchemy.model.patron import Loan
 from palace.manager.sqlalchemy.model.work import Work
 from palace.manager.util.datetime_helpers import utc_now
 from tests.fixtures.database import DatabaseTransactionFixture
-from tests.fixtures.files import FilesFixture
+from tests.fixtures.files import OPDS2FilesFixture
 
 
 class OPDS2Test:
@@ -89,19 +89,6 @@ class TestOPDS2ImporterFixture:
     data_source: DataSource
     importer: OPDS2Importer
     library: Library
-
-
-class OPDS2FilesFixture(FilesFixture):
-    """A fixture providing access to OPDS2 files."""
-
-    def __init__(self):
-        super().__init__("opds2")
-
-
-@pytest.fixture()
-def opds2_files_fixture() -> OPDS2FilesFixture:
-    """A fixture providing access to OPDS2 files."""
-    return OPDS2FilesFixture()
 
 
 @pytest.fixture
